requirements.txt
folium==0.14.0
inflection==0.5.1
numpy==1.21.5
pandas==1.4.2
Pillow==9.5.0
plotly==5.6.0
streamlit==1.21.0
streamlit_folium==0.11.1
